Move environs.Prepare* to environs/bootstrap
The environs.Prepare function is intrinsic to
bootstrapping, so we move it to the environs/bootstrap
package to keep bootstrap-related behaviour together.

This makes it possible for the PrepareParams struct
to contain bootstrap-specific information without
exposing that to the rest of "environs" and its many
consumers.
