+++ environs/open_test.go
@@ -9,8 +9,6 @@ import (
 	"github.com/juju/utils"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/cert"
-	"github.com/juju/juju/controller"
 	"github.com/juju/juju/environs"
 	"github.com/juju/juju/environs/bootstrap"
 	"github.com/juju/juju/environs/config"
@@ -53,7 +51,7 @@ func (s *OpenSuite) TestNewDummyEnviron(c *gc.C) {
 	ctx := envtesting.BootstrapContext(c)
 	cache := jujuclienttesting.NewMemStore()
 	controllerCfg := testing.FakeControllerBootstrapConfig()
-	env, err := environs.Prepare(ctx, cache, environs.PrepareParams{
+	env, err := bootstrap.Prepare(ctx, cache, bootstrap.PrepareParams{
 		ControllerConfig: controllerCfg,
 		ControllerName:   cfg.Name(),
 		BaseConfig:       cfg.AllAttrs(),
@@ -89,7 +87,7 @@ func (s *OpenSuite) TestUpdateEnvInfo(c *gc.C) {
 	c.Assert(err, jc.ErrorIsNil)
 	controllerCfg := testing.FakeControllerBootstrapConfig()
 	controllerCfg["controller-uuid"] = uuid
-	_, err = environs.Prepare(ctx, store, environs.PrepareParams{
+	_, err = bootstrap.Prepare(ctx, store, bootstrap.PrepareParams{
 		ControllerConfig: controllerCfg,
 		ControllerName:   "controller-name",
 		BaseConfig:       cfg.AllAttrs(),
@@ -134,145 +132,6 @@ func (*OpenSuite) TestNew(c *gc.C) {
 	c.Assert(err, gc.ErrorMatches, "model is not prepared")
 }
 
-func (*OpenSuite) TestPrepare(c *gc.C) {
-	baselineAttrs := dummy.SampleConfig().Merge(testing.Attrs{
-		"controller": false,
-		"name":       "erewhemos",
-	}).Delete(
-		"admin-secret",
-	)
-	cfg, err := config.New(config.NoDefaults, baselineAttrs)
-	c.Assert(err, jc.ErrorIsNil)
-	controllerStore := jujuclienttesting.NewMemStore()
-	ctx := envtesting.BootstrapContext(c)
-	controllerCfg := controller.Config{controller.ControllerUUIDKey: testing.ModelTag.Id()}
-	env, err := environs.Prepare(ctx, controllerStore, environs.PrepareParams{
-		ControllerConfig: controllerCfg,
-		ControllerName:   cfg.Name(),
-		BaseConfig:       cfg.AllAttrs(),
-		CloudName:        "dummy",
-	})
-	c.Assert(err, jc.ErrorIsNil)
-
-	// Check that an admin-secret was chosen.
-	adminSecret := env.Config().AdminSecret()
-	c.Assert(adminSecret, gc.HasLen, 32)
-	c.Assert(adminSecret, gc.Matches, "^[0-9a-f]*$")
-
-	// Check that the CA cert was generated.
-	cfgCertPEM, cfgCertOK := controllerCfg.CACert()
-	cfgKeyPEM, cfgKeyOK := controllerCfg.CAPrivateKey()
-	c.Assert(cfgCertOK, jc.IsTrue)
-	c.Assert(cfgKeyOK, jc.IsTrue)
-
-	// Check the common name of the generated cert
-	caCert, _, err := cert.ParseCertAndKey(cfgCertPEM, cfgKeyPEM)
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(caCert.Subject.CommonName, gc.Equals, `juju-generated CA for model "`+testing.SampleModelName+`"`)
-
-	// Check that controller was cached
-	foundController, err := controllerStore.ControllerByName(cfg.Name())
-	c.Assert(err, jc.ErrorIsNil)
-	c.Assert(foundController.ControllerUUID, gc.DeepEquals, cfg.UUID())
-	c.Assert(foundController.Cloud, gc.Equals, "dummy")
-
-	// Check we cannot call Prepare again.
-	env, err = environs.Prepare(ctx, controllerStore, environs.PrepareParams{
-		ControllerConfig: controllerCfg,
-		ControllerName:   cfg.Name(),
-		BaseConfig:       cfg.AllAttrs(),
-		CloudName:        "dummy",
-	})
-	c.Assert(err, jc.Satisfies, errors.IsAlreadyExists)
-	c.Assert(err, gc.ErrorMatches, `controller "erewhemos" already exists`)
-}
-
-func (*OpenSuite) TestPrepareGeneratesDifferentAdminSecrets(c *gc.C) {
-	baselineAttrs := dummy.SampleConfig().Merge(testing.Attrs{
-		"controller": false,
-		"name":       "erewhemos",
-	}).Delete(
-		"admin-secret",
-	)
-
-	ctx := envtesting.BootstrapContext(c)
-	env0, err := environs.Prepare(ctx, jujuclienttesting.NewMemStore(), environs.PrepareParams{
-		ControllerConfig: testing.FakeControllerBootstrapConfig(),
-		ControllerName:   "erewhemos",
-		BaseConfig:       baselineAttrs,
-		CloudName:        "dummy",
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	adminSecret0 := env0.Config().AdminSecret()
-	c.Assert(adminSecret0, gc.HasLen, 32)
-	c.Assert(adminSecret0, gc.Matches, "^[0-9a-f]*$")
-
-	// Allocate a new UUID, or we'll end up with the same config.
-	newUUID := utils.MustNewUUID()
-	baselineAttrs[config.UUIDKey] = newUUID.String()
-	controllerCfg := testing.FakeControllerBootstrapConfig()
-	controllerCfg[controller.ControllerUUIDKey] = newUUID.String()
-
-	env1, err := environs.Prepare(ctx, jujuclienttesting.NewMemStore(), environs.PrepareParams{
-		ControllerConfig: controllerCfg,
-		ControllerName:   "erewhemos",
-		BaseConfig:       baselineAttrs,
-		CloudName:        "dummy",
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	adminSecret1 := env1.Config().AdminSecret()
-	c.Assert(adminSecret1, gc.HasLen, 32)
-	c.Assert(adminSecret1, gc.Matches, "^[0-9a-f]*$")
-
-	c.Assert(adminSecret1, gc.Not(gc.Equals), adminSecret0)
-}
-
-func (*OpenSuite) TestPrepareWithMissingKey(c *gc.C) {
-	cfg, err := config.New(config.NoDefaults, dummy.SampleConfig().Merge(
-		testing.Attrs{
-			"controller": false,
-			"name":       "erewhemos",
-		},
-	))
-	c.Assert(err, jc.ErrorIsNil)
-	controllerStore := jujuclienttesting.NewMemStore()
-	controllerCfg := testing.FakeControllerBootstrapConfig()
-	delete(controllerCfg, "ca-private-key")
-	env, err := environs.Prepare(envtesting.BootstrapContext(c), controllerStore, environs.PrepareParams{
-		ControllerConfig: controllerCfg,
-		ControllerName:   cfg.Name(),
-		BaseConfig:       cfg.AllAttrs(),
-		CloudName:        "dummy",
-	})
-	c.Assert(err, gc.ErrorMatches, "cannot ensure CA certificate: controller configuration with a certificate but no CA private key")
-	c.Assert(env, gc.IsNil)
-}
-
-func (*OpenSuite) TestPrepareWithExistingKeyPair(c *gc.C) {
-	cfg, err := config.New(config.NoDefaults, dummy.SampleConfig().Merge(
-		testing.Attrs{
-			"controller": false,
-			"name":       "erewhemos",
-		},
-	))
-	c.Assert(err, jc.ErrorIsNil)
-	ctx := envtesting.BootstrapContext(c)
-	controllerCfg := testing.FakeControllerBootstrapConfig()
-	_, err = environs.Prepare(ctx, jujuclienttesting.NewMemStore(), environs.PrepareParams{
-		ControllerConfig: controllerCfg,
-		ControllerName:   cfg.Name(),
-		BaseConfig:       cfg.AllAttrs(),
-		CloudName:        "dummy",
-	})
-	c.Assert(err, jc.ErrorIsNil)
-	cfgCertPEM, cfgCertOK := controllerCfg.CACert()
-	cfgKeyPEM, cfgKeyOK := controllerCfg.CAPrivateKey()
-	c.Assert(cfgCertOK, jc.IsTrue)
-	c.Assert(cfgKeyOK, jc.IsTrue)
-	c.Assert(string(cfgCertPEM), gc.DeepEquals, testing.CACert)
-	c.Assert(string(cfgKeyPEM), gc.DeepEquals, testing.CAKey)
-}
-
 func (*OpenSuite) TestDestroy(c *gc.C) {
 	cfg, err := config.New(config.NoDefaults, dummy.SampleConfig().Merge(
 		testing.Attrs{
@@ -286,7 +145,7 @@ func (*OpenSuite) TestDestroy(c *gc.C) {
 	// the config storage info has been made.
 	controllerCfg := testing.FakeControllerBootstrapConfig()
 	ctx := envtesting.BootstrapContext(c)
-	e, err := environs.Prepare(ctx, store, environs.PrepareParams{
+	e, err := bootstrap.Prepare(ctx, store, bootstrap.PrepareParams{
 		ControllerConfig: controllerCfg,
 		ControllerName:   "controller-name",
 		BaseConfig:       cfg.AllAttrs(),
