+++ worker/logforwarder/orchestrator.go
@@ -7,7 +7,6 @@ import (
 	"github.com/juju/errors"
 
 	"github.com/juju/juju/api/base"
-	"github.com/juju/juju/environs"
 	"github.com/juju/juju/environs/config"
 )
 
@@ -37,9 +36,6 @@ type OrchestratorArgs struct {
 }
 
 func newOrchestratorForController(args OrchestratorArgs) (*orchestrator, error) {
-	if args.Config.Name() != environs.ControllerModelName {
-		return nil, errors.New("model-level log forwarding not supported")
-	}
 	controllerUUID := args.Config.UUID() // This won't work for per-model forwarding.
 
 	// For now we work with only 1 forwarder. Later we can have a proper
